==> all-features/3-extract.ipynb <==
# Assign positives and negatives 

import json
import threading
import concurrent.futures
import bz2
import csv
import time

import pandas
import py2neo
from tqdm import tqdm

## Connect to neo4j server

# Override the default py2neo timeout
py2neo.packages.httpstream.http.socket_timeout = 1e8

with open('servers.json') as read_file:
    instances = json.load(read_file)

name_to_neo = dict()
for instance in instances:
    uri = 'http://localhost:{}/db/data/'.format(instance['port'])
    bolt_port = 7690 + (instance['port'] % 10)
    name_to_neo[instance['name']] = py2neo.database.Graph(uri, bolt = True, bolt_port = bolt_port)

## Read metapaths

with open('data/metapaths.json') as read_file:
    metapaths = json.load(read_file)

metapaths.sort(key=lambda x: x['join_complexities'][0])
len(metapaths)

## Read hetnet-compound-disease pairs

part_df = pandas.read_table('data/partitions.tsv')
part_df['neo'] = part_df.hetnet.map(name_to_neo)
parts = list(part_df.itertuples())
print(len(part_df))
part_df.head(2)

part_df['disease_id'].nunique()

part_df['compound_id'].nunique()

## Set up queries

# Total number of queries
total_queries = len(metapaths) * len(part_df)
'{:,}'.format(total_queries)

parts = [row for row in part_df.itertuples()]

def generate_parameters(max_elems=None):
    """Generate compound, disease, metapath combinations"""
    n = 0
    for metapath_dict in metapaths:
        metapath = metapath_dict['abbreviation']
        query = metapath_dict['dwpc_query']
        for part_info in parts:
            if max_elems is not None and n == max_elems:
                break
            yield {
                'neo': part_info.neo,
                'hetnet': part_info.hetnet,
                'compound_id': part_info.compound_id,
                'disease_id': part_info.disease_id,
                'metapath': metapath,
                'query': query,
                'w': 0.4,
            }
            n += 1

def compute_dwpc(neo, hetnet, query, metapath, compound_id, disease_id, w):
    """Execute the neo4j query and write results to file"""
    start = time.time()
    results = neo.data(query, source=compound_id, target=disease_id, w=w)
    record = results[0]
    seconds = '{0:.4g}'.format(time.time() - start)
    row = hetnet, compound_id, disease_id, metapath, record['PC'], w, '{0:.6g}'.format(record['DWPC']), seconds
    with writer_lock:
        writer.writerow(row)

## Execute queries

def test(neo, hetnet, query, metapath, compound_id, disease_id, w):
    print(neo, hetnet, query, metapath, compound_id, disease_id, w)

compute_dwpc(**params)

query = 'MATCH (a:Compound)-[:`drug-used-for-treatment_DduftC`]-(other) RETURN a.name, count(other) LIMIT 4'

g.data(query)

test(**params)

metapath_dict['abbreviation']

for metapath_dict in metapaths:
    metapath = metapath_dict['abbreviation']
    query = metapath_dict['dwpc_query']
    part_info = parts[1854]
    params = {
                    'neo': part_info.neo,
                    'hetnet': part_info.hetnet,
                    'compound_id': part_info.compound_id,
                    'disease_id': part_info.disease_id,
                    'metapath': metapath,
                    'query': query,
                    'w': 0.4,
                }



    g = params['neo']
    query = params['query']
    source = params['compound_id']
    target = params['disease_id']
    w = params['w']

    res = g.data(query, source=source, target=target, w=w)
    if res[0]['DWPC'] != 0:
        print(res)


%%time

# Parameters
workers = 4
max_elems = None

# Prepare writer
print('preparing writer...')
path = 'data/dwpc.tsv.bz2'
write_file = bz2.open(path, 'wt')
writer = csv.writer(write_file, delimiter='\t')
writer.writerow(['hetnet', 'compound_id', 'disease_id', 'metapath', 'PC', 'w', 'DWPC', 'seconds'])

# Create ThreadPoolExecutor
print('creating executer...')
executor = concurrent.futures.ThreadPoolExecutor(max_workers=workers)
writer_lock = threading.Lock()

# Submit jobs
print('submitting first job...')
n_queries = 0
for params in tqdm(generate_parameters(max_elems)):
    executor.submit(compute_dwpc, **params)
    n_queries += 1

# Shutdown and close
executor.shutdown()
write_file.close()

n_queries
